Separate test file for baseEcAdapter instances
